Add files via upload
The client could take command line arguments specifying a server name or IP address, the port on which to contact the server, the method, and the path of the requested object on the server. I have implemented two methods of HTTP: GET and PUT.
·GET
The format of the command line is: myclient host port_number GET filename
The basic client GET action proceeds as follows:
(1)Connect to the server via a connection-oriented socket.
(2) Submit a valid HTTP/1.0 GET request for the supplied URL.
(3) Read (from the socket) the server's response and display it as program output.
·PUT
The format of the command line is: myclient host port_number PUT filename
The basic client PUT action proceeds as follows:
(1) Connect to the server via a connection-oriented socket.
(2) Submit a PUT request for the supplied file.
(3) Send the file to the server.
(4) Wait for server's reply.